Initial version adding eMMC support
This commit creates a rev D board with an optional eMMC variant for
storage.

Some notes:

The eMMC chip generally handles the state internally when enabling
4 or 8 bit data modes. So having the spare data lines connected to
NDQ[4:7] should be no problem. This needs to be tested however and
with some soldering skills should be possible to test it by soldering
to the balls directly.

The only real issue is the stage 0 bootloader (in the SoC itself) as any
moment after that, we can ground the Hi-Z the pins from the SPL and
onwards. We might even get lucky and get 8 bit data transfer mode, since
the silicon and the driver support 8 bit mode, we just don't have any
muxes exposed according to the docs.

Additionally, the capacitors for the eMMC power supply should be the
same as the NAND flash uses, but I haven't figured out how to do this in
eagle.

The VCCQ, VDDI capacitors could always be populated, the'd end up being
unused, making the eMMC or NAND option a hot-swap thing, where either
can be used.

Finally, the nRST pin can be left unconnected thus leaving the 2
resistors off entirely. I would still route them though, as in theory,
we can then use the GPIO as a hardware reset for the eMMC if ever
needed. Because of this though, i'd leave them unpopulated.

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>